drop is working

+++ src/commanders/base_commander.py
@@ -33,9 +33,7 @@ class BaseCommander(abc.ABC):
         pass
 
     @abc.abstractmethod
-    async def goto_position(
-        self, latitude: float, longitude: float, altitude: float
-    ) -> None:
+    async def goto_position(self, latitude: float, longitude: float, altitude: float) -> None:
         """
         Command drone to fly to specified GPS position.
         """
@@ -61,9 +59,7 @@ class BaseCommander(abc.ABC):
         pass
 
     @abc.abstractmethod
-    async def set_pcmds(
-        self, roll: float, pitch: float, yaw: float, gaz: float
-    ) -> None:
+    async def set_pcmds(self, roll: int, pitch: int, yaw: int, gaz: int) -> None:
         """
         Set the drone's pitch/roll/yaw/gaz commands.
         """
